Apply existing protobuf buffer optimization to nested message encoding (~2.3x speed up) (#9458)
